--- src/main/java/io/javabrains/CourseApiDataApplication.java
@@ -1,18 +1,0 @@
-package io.javabrains;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceAutoConfiguration;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@SpringBootApplication(exclude = {DataSourceAutoConfiguration.class })
-//@SpringBootApplication
-public class CourseApiDataApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CourseApiDataApplication.class, args);
-    }
-
-}
