+++ src/main/java/io/javabrains/springbootstarter/topic/CourseApiDataApplication.java
@@ -1,0 +1,11 @@
+package io.javabrains.springbootstarter.topic;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication()
+public class CourseApiDataApplication {
+    public static void main(String[] args) {
+        SpringApplication.run(CourseApiDataApplication.class, args);
+    }
+}
