--- src/main/java/io/javabrains/springbootstarter/topic/TopicController.java
@@ -1,57 +1,0 @@
-package io.javabrains.springbootstarter.topic;
-
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.lang.NonNull;
-import org.springframework.web.bind.annotation.DeleteMapping;
-import org.springframework.web.bind.annotation.GetMapping;
-import org.springframework.web.bind.annotation.PathVariable;
-import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestMethod;
-import org.springframework.web.bind.annotation.RestController;
-
-import java.util.List;
-
-@RestController
-public class TopicController {
-
-    private final TopicService topicService;
-
-    /*
-    @Autowired makes Spring look in registry,
-    sees that a TopicService @Service instance already exists, and injects it.
-     */
-    @Autowired
-    public TopicController(@NonNull TopicService topicService) {
-        this.topicService = topicService;
-    }
-
-    @GetMapping("/topics")
-    public List<Topic> getAllTopics() {
-        return topicService.getAllTopics();
-    }
-
-    @RequestMapping("/topics/{id}")
-    @GetMapping("/topics/{id}")
-    public Topic getTopic(@PathVariable String id) {
-        return topicService.getTopic(id);
-    }
-
-    @RequestMapping(value = "/topics", method = RequestMethod.POST)
-    @PostMapping("/topics")
-    public void addTopic(@RequestBody Topic topic) {
-        topicService.addTopic(topic);
-    }
-
-    @PutMapping("/topics/{id}")
-    public void updateTopic(@PathVariable String id, @RequestBody Topic topic) {
-        topicService.updateTopic(id, topic);
-    }
-
-    @DeleteMapping("/topics/{id}")
-    public void deleteTopic(@PathVariable String id) {
-        topicService.deleteTopic(id);
-    }
-}
