+++ src/main/resources/application.properties
@@ -1,1 +1,5 @@
-#spring.datasource.url=jdbc:derby:memory:local;create=true
+spring.datasource.url=jdbc:h2:mem:testdb
+spring.datasource.driverClassName=org.h2.Driver
+spring.datasource.username=sa
+spring.datasource.password=password
+spring.jpa.database-platform=org.hibernate.dialect.H2Dialect
